requirements.txt
sentence-transformers==2.7.0
openai==1.30.0
google-generativeai==0.5.4
python-dotenv==1.0.1
numpy==1.26.4
pytest==8.2.0
jupyter==1.0.0
notebook==7.2.0
matplotlib==3.9.0
seaborn==0.13.2
